requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
pytest==8.0.0
python-dotenv==1.1.0
httpx==0.26.0  # For async HTTP client testing
pyarrow==19.0.1
pytest-asyncio==0.23.5  # For async test support
pytest-cov==4.1.0  # For coverage reporting
pandas==2.2.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
